+++ lib/utils.dart
@@ -4,5 +4,6 @@ library utils;
 export 'src/utils/api_error.dart';
 export 'src/utils/api_exception.dart';
 export 'src/utils/encoding.dart';
+export 'src/utils/mime.dart';
 export 'src/utils/token.dart';
 export 'src/utils/utils.dart';